投稿レビューの更新、削除の権限による出しわけ処理を追記

+++ app/api/review/delete/[id]/route.tsx
@@ -6,11 +6,17 @@ export async function DELETE(
   request: NextRequest,
   context: { params: Promise<{ id: string }> }
 ) {
+  const reqBody = request.json();
   try {
     await conntectDB();
     const { id } = await context.params;
-    await ReviewModel.deleteOne({ _id: id });
-    return NextResponse.json({ message: "レビュー削除成功" });
+    const singleReview = await ReviewModel.findById(id);
+    if (singleReview.email === reqBody.email) {
+      await ReviewModel.deleteOne({ _id: id });
+      return NextResponse.json({ message: "レビュー削除成功" });
+    } else {
+      return NextResponse.json({ message: "他の人が投稿したレビューです" });
+    }
   } catch {
     return NextResponse.json({ message: "レビュー削除失敗" });
   }
